Reworking PMM init some

+++ yuki/inc/mm/vmm.hpp
@@ -12,6 +12,6 @@ typedef struct
 uint64_t createPte(uint64_t physicalAddr, uint64_t flags);
 void initVmm(limine_memmap_response *memoryMap, std::uint64_t hhdm);
 void mapPage(uint64_t virtualAddr, uint64_t physicalAddr, uint64_t flags);
-void mapPages(uint64_t virtualStart, uint64_t physicalStart, uint8_t flags, uint64_t count);
+void mapPages(uint64_t virtualStart, uint64_t physicalStart, uint64_t flags, uint64_t count);
 void unmapPage(uint64_t virtualAddr);
 void remapPage(uint64_t virtualAddr);
